Update README.md
# multimedia-information-retrieval

The goal of this project is to effectively perform classification, clustering, and (graph) node ranking tasks on a set of images, which we achieve through the following algorithms: K-Nearest Neighbours, PageRank, Personalized PageRank, Spectral Partitioning, and Min-a-Max. Also, we implement an indexing technique, Locality Sensitive Hashing, to target the high dimensional data presented in the form of visual descriptors.